css/css-values/minmax-number-serialize.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-values-4/#comp-func">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#numbers">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#calc-serialize">
<link rel="author" title="Xiaocheng Hu" href="mailto:xiaochengh@chromium.org">
<link rel="author" title="Tab Atkins-Bittner" href="https://xanthir.com/contact">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/serialize-testcommon.js"></script>
<div id=target></div>
<script>
function test_serialization(t,s,c,u, {prop="opacity"}={}) {
    test_specified_serialization(prop, t, s);
    test_computed_serialization(prop, t, c);
    if(u) test_used_serialization(prop, t, u);
}

test_serialization(
    'min(.1)',
    'calc(0.1)',
    '0.1');
test_serialization(
    'max(.1)',
    'calc(0.1)',
    '0.1');

test_serialization(
    'min(.1, .2, .3)',
    'calc(0.1)',
    '0.1');
test_serialization(
    'max(.1, .2, .3)',
    'calc(0.3)',
    '0.3');

test_serialization(
    'min(.3, .2, .1)',
    'calc(0.1)',
    '0.1');
test_serialization(
    'max(.3, .2, .1)',
    'calc(0.3)',
    '0.3');

test_serialization(
    'calc(min(.1) + min(.2))',
    'calc(0.3)',
    '0.3');
test_serialization(
    'calc(max(.1) + max(.2))',
    'calc(0.3)',
    '0.3');

test_serialization(
    'calc(.1 + min(.1))',
    'calc(0.2)',
    '0.2');
test_serialization(
    'calc(max(.1) + .1)',
    'calc(0.2)',
    '0.2');
</script>
